\name{ecospat.migclim}

\alias{ecospat.migclim}

\title{Implementing dispersal into species distribution models}

\description{Enables the implementation of species-specific dispersal constraints into projections of species distribution models under environmental change and/or landscape fragmentation scenarios.}

\usage{ecospat.migclim ()}

\details{The MigClim model is a cellular automaton originally designed to implement dispersal constraints into projections of species distributions under environmental change and landscape fragmentation scenarios.}

\references{Engler, R., W. Hordijk and A. Guisan. 2012. The MIGCLIM R package -- seamless integration of dispersal constraints into projections of species distribution models. \emph{Ecography}, \bold{35}, 872-878.

Engler, R. and A. Guisan. 2009. MIGCLIM: predicting plant distribution and dispersal in a changing climate. \emph{Diversity and Distributions}, \bold{15}, 590-601.

Engler, R., C.F. Randin, P. Vittoz, T. Czaka, M. Beniston, N.E. Zimmermann and A. Guisan. 2009. Predicting future distributions of mountain plants under climate change: does dispersal capacity matter? \emph{Ecography}, \bold{32}, 34-45.
}

\author{Robin Engler \email{robin.engler@gmail.com}, Wim Hordijk \email{wim@WorldWideWanderings.net} and Loic Pellissier \email{loic.pellissier@unifr.ch}}

\examples{
ecospat.migclim ()
}