\name{ecospat-package}
\alias{ecospat-package}
\alias{ecospat}

\docType{package}

\title{Spatial Ecology Miscellaneous Methods}
\description{
Collection of R functions and data sets for the support of spatial ecology analyses with a focus on pre-, core and post- modelling analyses of species distribution, niche quantification and community assembly., written by current and former members and collaborators of the ecospat group of Antoine Guisan, Department of Ecology and Evolution (DEE) & Institute of Earth Surface Dynamics (IDYST), University of Lausanne, Switzerland.
}
\details{
\tabular{ll}{
Package: \tab ecospat\cr
Type: \tab Package\cr
Version: \tab 2.0.1\cr
Date: \tab 2016-06-09\cr
License: \tab GPL \cr
}

}
\author{Olivier Broennimann [aut],
             Valeria Di Cola [cre, aut],
             Blaise Petitpierre [ctb],
             Frank Breiner [ctb],
             Manuela D`Amen [ctb],
             Christophe Randin [ctb],
             Robin Engler [ctb],
             Wim Hordijk [ctb],
             Julien Pottier [ctb],
             Mirko Di Febbraro [ctb],
             Loic Pellissier [ctb],
             Dorothea Pio [ctb],
             Ruben Garcia Mateo [ctb],
             Anne Dubuis [ctb],
             Daniel Scherrer [ctb],
             Luigi Maiorano [ctb],
             Achilleas Psomas [ctb],
             Charlotte Ndiribe [ctb]
             Nicolas Salamin [ctb],
             Niklaus Zimmermann [ctb],
             Antoine Guisan [aut]
             }

\keyword{ package }
